requirements.txt
aiohttp==3.10.11
discord.py==2.6.3
psycopg[binary]==3.2.10
pytest==8.3.5
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pytz==2024.2
